requirements.txt
aiohttp>=2.2.5
async-timeout>=2.0.1
attrs>=17.4.0
chardet>=3.0.4
discord.py>=1.0.0
diskcache>=3.0.1
idna>=2.6
idna-ssl>=1.0.1
multidict>=4.1.0
websockets>=4.0.1
yarl>=1.1.1
python-dateutil>=0.1
